Add a few more tests and put v1.0.0 tests in their own module
