+++ ckanext/versioned_datastore/tests/test_multisearch.py
@@ -3,7 +3,6 @@ from collections import OrderedDict
 
 import jsonschema
 import os
-
 from ckanext.versioned_datastore.lib.multisearch import load_schemas, get_latest_query_version, \
     validate_query, InvalidQuerySchemaVersionError, translate_query
 from ckantest.models import TestBase
@@ -77,222 +76,3 @@ class TestMultisearch(TestBase):
         with patch(u'ckanext.versioned_datastore.lib.multisearch.schemas', {}):
             with assert_raises(InvalidQuerySchemaVersionError):
                 translate_query(MagicMock(), MagicMock())
-
-
-class TestV1_0_0Translator(TestBase):
-    plugins = [u'versioned_datastore']
-
-    @staticmethod
-    def compare_query_and_search(query, search_dict):
-        assert_equal(translate_query(query, u'v1.0.0').to_dict(), search_dict)
-
-    def test_translate_1(self):
-        query = {}
-        search_dict = {}
-        self.compare_query_and_search(query, search_dict)
-
-    def test_translate_2(self):
-        query = {
-            u'search': u'mollusca'
-        }
-        search_dict = {
-            u'query': {
-                u'match': {
-                    u'meta.all': {
-                        u'query': u'mollusca',
-                        u'operator': u'and'
-                    }
-                }
-            }
-        }
-        self.compare_query_and_search(query, search_dict)
-
-    def test_translate_3(self):
-        query = {
-            u"filters": {
-                u"and": [
-                    {
-                        u"string_equals": {
-                            u"fields": [
-                                u"genus"
-                            ],
-                            u"value": u"helix"
-                        }
-                    },
-                    {
-                        u"string_contains": {
-                            u"fields": [
-                                u"higherGeography"
-                            ],
-                            u"value": u"europe"
-                        }
-                    }
-                ]
-            }
-        }
-        search_dict = {
-            u'query': {
-                u'bool': {
-                    u'must': [
-                        {
-                            u'term': {
-                                u'data.genus': u'helix'
-                            }
-                        },
-                        {
-                            u'match': {
-                                u'data.higherGeography.full': {
-                                    u'query': u'europe',
-                                    u'operator': u'and'
-                                }
-                            }
-                        }
-                    ]
-                }
-            }
-        }
-        self.compare_query_and_search(query, search_dict)
-
-    def test_translate_4(self):
-        query = {
-            u'search': u'italy',
-            u"filters": {
-                u"and": [
-                    {
-                        u"string_equals": {
-                            u"fields": [
-                                u"genus"
-                            ],
-                            u"value": u"helix"
-                        }
-                    },
-                    {
-                        u"string_contains": {
-                            u"fields": [
-                                u"higherGeography"
-                            ],
-                            u"value": u"europe"
-                        }
-                    }
-                ]
-            }
-        }
-        search_dict = {
-            u'query': {
-                u'bool': {
-                    u'must': [
-                        {
-                            u'term': {
-                                u'data.genus': u'helix'
-                            }
-                        },
-                        {
-                            u'match': {
-                                u'data.higherGeography.full': {
-                                    u'query': u'europe',
-                                    u'operator': u'and'
-                                }
-                            }
-                        },
-                        {
-                            u'match': {
-                                u'meta.all': {
-                                    u'query': u'italy',
-                                    u'operator': u'and'
-                                }
-                            }
-                        }
-                    ]
-                }
-            }
-        }
-        self.compare_query_and_search(query, search_dict)
-
-    def test_translate_5(self):
-        query = {
-            u"filters": {
-                u"and": [
-                    {
-                        u"string_equals": {
-                            u"fields": [
-                                u"genus"
-                            ],
-                            u"value": u"helix"
-                        }
-                    },
-                    {
-                        u"or": [
-                            {
-                                u"string_contains": {
-                                    u"fields": [
-                                        u"higherGeography"
-                                    ],
-                                    u"value": u"italy"
-                                }
-                            },
-                            {
-                                u"string_contains": {
-                                    u"fields": [
-                                        u"higherGeography"
-                                    ],
-                                    u"value": u"spain"
-                                }
-                            },
-                            {
-                                u"string_contains": {
-                                    u"fields": [
-                                        u"higherGeography"
-                                    ],
-                                    u"value": u"portugal"
-                                }
-                            }
-                        ]
-                    }
-                ]
-            }
-        }
-        search_dict = {
-            u'query': {
-                u'bool': {
-                    u'must': [
-                        {
-                            u'term': {
-                                u'data.genus': u'helix'
-                            }
-                        },
-                        {
-                            u'bool': {
-                                u'minimum_should_match': 1,
-                                u'should': [
-                                    {
-                                        u'match': {
-                                            u'data.higherGeography.full': {
-                                                u'query': u'italy',
-                                                u'operator': u'and'
-                                            }
-                                        }
-                                    },
-                                    {
-                                        u'match': {
-                                            u'data.higherGeography.full': {
-                                                u'query': u'spain',
-                                                u'operator': u'and'
-                                            }
-                                        }
-                                    },
-                                    {
-                                        u'match': {
-                                            u'data.higherGeography.full': {
-                                                u'query': u'portugal',
-                                                u'operator': u'and'
-                                            }
-                                        }
-                                    }
-                                ]
-                            }
-                        }
-                    ]
-                }
-            }
-        }
-        self.compare_query_and_search(query, search_dict)
